LPS-57312 move
