+++ modules/frontend/frontend-taglib-web/src/com/liferay/frontend/taglib/servlet/ServletContextUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.util;
+package com.liferay.frontend.taglib.servlet;
 
 import javax.servlet.ServletContext;
